refactor: move finding.go from handler.go
