+++ pkg/scc/finding_test.go
